if statment added.
wheneverthe ball moves past the 0 of the y axis the speed of the paddles move by .1 everytime making the game fun && exciting ;)

+++ exercises/Final/Paddles.pde
@@ -3,7 +3,7 @@ class Paddles {
 
   int x;
   int y;
-  int vy = 1;
+  float vy = 1;
   float vx;
 
   int size;
